Adding template to be used with IEEETran.cls
The IEEETrans class can be found at

  http://www.michaelshell.org/tex/ieeetran/

The template modifies minimally the default latex template:

  * fixes the document class to `IEEETran`
  * uses custom `IEEEauthorblock*` commands for the authors section

The variables expected for the author metadata (in YAML format) are of
the form:

    title: 'A Title'
    author:
    - names: Student Foo, Advisor Bar
      affiliation: University of Foo
    - names: Other Collaborator
      affiliation: Collaboration Agency
    - names: A Third One
      affiliation: Coolest Place

The above assumes that at most three affiliations will be there. If not,
then the alternative format has to be used but that's not supported
currently.